+++ ee-express-core/src/test/java/de/simonkerstan/ee/core/configuration/types/TestEnum.java
@@ -1,0 +1,17 @@
+/*
+ * Copyright (c) 2025 Simon Kerstan
+ * SPDX-License-Identifier: MIT
+ */
+
+package de.simonkerstan.ee.core.configuration.types;
+
+/**
+ * Test enum for testing the configuration property resolver.
+ */
+public enum TestEnum {
+
+    X,
+
+    Y
+
+}