+++ ee-express-core/src/test/java/de/simonkerstan/ee/core/modules/FrameworkModuleLoaderUnitTest.java
@@ -1,12 +1,19 @@
+/*
+ * Copyright (c) 2025 Simon Kerstan
+ * SPDX-License-Identifier: MIT
+ */
+
 package de.simonkerstan.ee.core.modules;
 
 import de.simonkerstan.ee.core.CoreModule;
 import de.simonkerstan.ee.core.classpath.ClasspathResolver;
+import de.simonkerstan.ee.core.modules.test.AfterModule;
 import de.simonkerstan.ee.core.modules.test.TestModule;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertSame;
 
 class FrameworkModuleLoaderUnitTest {
 
@@ -14,13 +21,15 @@ class FrameworkModuleLoaderUnitTest {
     @DisplayName("Load all modules -> Should load all modules")
     void testLoadModules() {
         final var modules = FrameworkModuleLoader.loadFrameworkModules(ClasspathResolver.getWrapperForFullClasspath());
-        assertEquals(2, modules.size());
+        assertEquals(3, modules.size());
 
-        // Assert TestModule and CoreModule is loaded (0 = core module; 1 = test module)
+        // Assert TestModule and CoreModule is loaded (0 = core module; 1 = test module; 2 = after module)
         final var coreModule = modules.get(0);
         assertSame(CoreModule.class, coreModule.getClass());
         final var testModule = modules.get(1);
         assertSame(TestModule.class, testModule.getClass());
+        final var afterModule = modules.get(2);
+        assertSame(AfterModule.class, afterModule.getClass());
     }
 
 }