+++ ee-express-persistence/src/main/java/de/simonkerstan/ee/persistence/configuration/SchemaStrategy.java
@@ -1,0 +1,15 @@
+/*
+ * Copyright (c) 2025 Simon Kerstan
+ * SPDX-License-Identifier: MIT
+ */
+
+package de.simonkerstan.ee.persistence.configuration;
+
+/**
+ * Strategy for handling the database schema (only for relational persistence).
+ */
+public enum SchemaStrategy {
+
+    CREATE
+
+}